Add shfmt and python linters
Will be adding more linters for each language as it comes up.

`run_with_args` is a custom rule that captures the tool and arguments
via a template. This is to get around a limitation of bazel where
arguments to rules are lost if that rule is a dependency of another
rule. The example is `tools/lint:lint` where it runs each linter
serially without having to know all the arguments.