rules_oci working similar to rules_docker
Add 2 example commands - go and python. The go command is loaded onto
the nginx container just to demonstrate how to do it. The python command
will be eventually added alongside deps.

Add an `image` macro that defines a few other pieces that were present
in rules_docker: running the image and producing a digest.